requirements.txt
PyYAML~=6.0
requests~=2.28.1
APScheduler~=3.6.3
zhconv~=1.4.3
python-telegram-bot~=13.14
sqlalchemy~=1.4.44
pillow~=9.3.0